Improve publishing process
There was a bug where publishing was broken if a part
of the version was larger than 9.

Instead of using custom actions, we are now using
EndBug/version-check@v2 to detect that the version
has changed.

Since debugging this was painful, we also started
caching node_modules keying on the sha so we
don't run npm ci on every job.

Signed-off-by: Juan Mendes <mendejuan@gmail.com>