Avoid using `;` and commands not supported by Windows
Using `cmdA && cmdB` means instead of `;` not only
prevents `cmdB` from executing if cmdA fails, it also
breaks npm scripts on windows unless you make sure the
semi-colon is separated by spaces, which is not the norm.

Following commands were not supported on Windows
* `mkdir -p`, adding [mkdirp](https://www.npmjs.com/package/mkdirp)
* `rsync and cp`, adding [copfyfiles](https://www.npmjs.com/package/copyfiles).

I initially used `cpx` but it does not work from an npm script on Windows. That
is, it did not copy any files at all. It was striking that it did work if I ran
`npx cpx` directly from the command line.

Used cross-spawn because spawn wasn't working on windows.
Signed-off-by: Juan Mendes <mendejuan@gmail.com>